Creating the big-bang game
This commit adds the big-bang game and unit tests
for interview purposes.

The factory design pattern has been used for this
implementation in order to create different types of
Choices (e.g. Paper, Rock, etc). Using that pattern,
the design is reusable and extensible without
repetitions. The rules of the game as described in
the exercise have been applied into the code.

Run it using "python3 paper_rock.py".
Stop it with 'Ctrl-C'.

Run unit tests using "pytest test.py"